// src/api/orval/model/performancePerformanceDbQueriesResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */
import type { PerformancePerformanceDbQueryEntryDto } from './performancePerformanceDbQueryEntryDto';

export interface PerformancePerformanceDbQueriesResponse {
  /** @nullable */
  items?: PerformancePerformanceDbQueryEntryDto[] | null;
}