// src/api/orval/model/performancePerformanceDbQueryEntryDto.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */

export interface PerformancePerformanceDbQueryEntryDto {
  timestampUtc?: string;
  /** @nullable */
  requestId?: string | null;
  durationMs?: number;
  /** @nullable */
  commandType?: string | null;
  /** @nullable */
  sql?: string | null;
  succeeded?: boolean;
  /** @nullable */
  exceptionMessage?: string | null;
}