// src/api/orval/model/performancePerformanceHttpRequestEntryDto.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */

export interface PerformancePerformanceHttpRequestEntryDto {
  timestampUtc?: string;
  /** @nullable */
  requestId?: string | null;
  /** @nullable */
  method?: string | null;
  /** @nullable */
  path?: string | null;
  statusCode?: number;
  durationMs?: number;
  /** @nullable */
  userName?: string | null;
  /** @nullable */
  traceId?: string | null;
}