// src/api/orval/model/performancePerformanceHttpRequestsResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */
import type { PerformancePerformanceHttpRequestEntryDto } from './performancePerformanceHttpRequestEntryDto';

export interface PerformancePerformanceHttpRequestsResponse {
  /** @nullable */
  items?: PerformancePerformanceHttpRequestEntryDto[] | null;
}